extract vm to own class
